norminette corrections

+++ src/path.c
@@ -6,7 +6,7 @@
 /*   By: pedro <pedro@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/26 14:16:46 by pmessett          #+#    #+#             */
-/*   Updated: 2023/06/14 14:49:13 by pedro            ###   ########.fr       */
+/*   Updated: 2023/06/15 19:15:41 by pedro            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -62,57 +62,55 @@ t_path	*find_path(t_path *path_list, char **possible_paths, char *av,
 	return (path_list);
 }
 
-/*Try to acess the path passes as argument*/
-t_path	*absolute_path(t_path *path_list, char *av, char **path_and_cmd)
+t_path	*handle_str_error(char **av, int i, int ac, t_path *path_list)
 {
-	char	*path;
+	ft_printf("pipex: %s: command not found\n", av[i]);
+	if (i == ac - 2 && !is_here_doc(av[1]))
+	{
+		free_path_list(&path_list);
+		unlink(".tmp");
+		exit(EXIT_FAILURE);
+	}
+	path_list = set_path_list(path_list, NULL, NULL);
+	return (path_list);
+}
 
-	path = ft_strdup(av);
-	if (access(av, F_OK) == -1)
-		ft_printf("pipex: %s: No such file or directory\n", path_and_cmd[0]);
-	free(path_and_cmd[0]);
-	path_and_cmd[0] = path;
-	path_and_cmd[1] = NULL;
-	path_list = set_path_list(path_list, path_and_cmd[0], path_and_cmd);
+t_path	*choose_handle(char *av, char **envp, t_path *path_list)
+{
+	char	**path_and_cmd;
+	char	**possible_paths;
+
+	path_and_cmd = ft_split(av, ' ');
+	if (!(ft_strncmp(av, "/", 1)) || !(ft_strncmp(av, "./", 2)))
+	{
+		if (access(path_and_cmd[0], F_OK) == -1)
+			ft_printf("pipex: %s: No such file or directory\n",
+				path_and_cmd[0]);
+		path_list = set_path_list(path_list, path_and_cmd[0], path_and_cmd);
+	}
+	else
+	{
+		possible_paths = set_possible_paths(envp);
+		path_list = find_path(path_list, possible_paths, av, path_and_cmd);
+		free_matrix(possible_paths);
+	}
 	return (path_list);
 }
 
 /*Main function to define if the argv is an absolute path or not*/
 t_path	*define_path(t_path *path_list, int ac, char **av, char **envp)
 {
-	char	**path_and_cmd;
-	char	**possible_paths;
-	int		i;
+	int	i;
 
 	i = 1;
 	if (is_here_doc(av[1]))
 		i = 2;
 	while (av[++i] && i < ac - 1)
 	{
-		if (!*av[i] || ft_str_is_space(av[i]))
-		{
-			ft_printf("pipex: %s: command not found\n", av[i]);
-			if (i == ac - 2)
-			{
-				free_path_list(&path_list);
-				unlink(".tmp");
-				exit(EXIT_FAILURE);
-			}
-			path_list = set_path_list(path_list, NULL, NULL);
-		}
+		if (ft_str_error(av[i]))
+			path_list = handle_str_error(av, i, ac, path_list);
 		else
-		{
-			path_and_cmd = ft_split(av[i], ' ');
-			if (!(ft_strncmp(av[i], "/", 1)) || !(ft_strncmp(av[i], "./", 2)))
-				path_list = absolute_path(path_list, av[i], path_and_cmd);
-			else
-			{
-				possible_paths = set_possible_paths(envp);
-				path_list = find_path(path_list, possible_paths, av[i],
-					path_and_cmd);
-				free_matrix(possible_paths);
-			}
-		}
+			path_list = choose_handle(av[i], envp, path_list);
 	}
 	return (path_list);
 }
